requirements.txt
opencv-python
Pillow
tqdm
tensorflow
imutils
matplotlib